Moved `CoreDataFeedMapper` to separate file
